+++ homeassistant/components/radarr/__init__.py
@@ -2,12 +2,11 @@
 
 from __future__ import annotations
 
-from dataclasses import dataclass, fields
+from dataclasses import fields
 
 from aiopyarr.models.host_configuration import PyArrHostConfiguration
 from aiopyarr.radarr_client import RadarrClient
 
-from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import CONF_API_KEY, CONF_URL, CONF_VERIFY_SSL, Platform
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
@@ -18,24 +17,13 @@ from .coordinator import (
     HealthDataUpdateCoordinator,
     MoviesDataUpdateCoordinator,
     QueueDataUpdateCoordinator,
+    RadarrConfigEntry,
+    RadarrData,
     RadarrDataUpdateCoordinator,
     StatusDataUpdateCoordinator,
 )
 
 PLATFORMS = [Platform.BINARY_SENSOR, Platform.CALENDAR, Platform.SENSOR]
-type RadarrConfigEntry = ConfigEntry[RadarrData]
-
-
-@dataclass(kw_only=True, slots=True)
-class RadarrData:
-    """Radarr data type."""
-
-    calendar: CalendarUpdateCoordinator
-    disk_space: DiskSpaceDataUpdateCoordinator
-    health: HealthDataUpdateCoordinator
-    movie: MoviesDataUpdateCoordinator
-    queue: QueueDataUpdateCoordinator
-    status: StatusDataUpdateCoordinator
 
 
 async def async_setup_entry(hass: HomeAssistant, entry: RadarrConfigEntry) -> bool:
@@ -50,12 +38,14 @@ async def async_setup_entry(hass: HomeAssistant, entry: RadarrConfigEntry) -> bo
         session=async_get_clientsession(hass, entry.data[CONF_VERIFY_SSL]),
     )
     data = RadarrData(
-        calendar=CalendarUpdateCoordinator(hass, host_configuration, radarr),
-        disk_space=DiskSpaceDataUpdateCoordinator(hass, host_configuration, radarr),
-        health=HealthDataUpdateCoordinator(hass, host_configuration, radarr),
-        movie=MoviesDataUpdateCoordinator(hass, host_configuration, radarr),
-        queue=QueueDataUpdateCoordinator(hass, host_configuration, radarr),
-        status=StatusDataUpdateCoordinator(hass, host_configuration, radarr),
+        calendar=CalendarUpdateCoordinator(hass, entry, host_configuration, radarr),
+        disk_space=DiskSpaceDataUpdateCoordinator(
+            hass, entry, host_configuration, radarr
+        ),
+        health=HealthDataUpdateCoordinator(hass, entry, host_configuration, radarr),
+        movie=MoviesDataUpdateCoordinator(hass, entry, host_configuration, radarr),
+        queue=QueueDataUpdateCoordinator(hass, entry, host_configuration, radarr),
+        status=StatusDataUpdateCoordinator(hass, entry, host_configuration, radarr),
     )
     for field in fields(data):
         coordinator: RadarrDataUpdateCoordinator = getattr(data, field.name)
